Code hygiene: split workflow steps into separate files
