+++ src/mastra/workflows/schemas/index.ts
@@ -1,0 +1,4 @@
+// Export all schemas from a central location
+export { frustrationsSchema } from './frustrations';
+export { memeTemplateSchema } from './meme-template';
+export { captionsSchema } from './captions';
