+++ src/mastra/workflows/steps/find-base-meme.ts
@@ -1,28 +1,29 @@
-import { createTool } from '@mastra/core';
+import { createStep } from '@mastra/core/workflows';
 import { z } from 'zod';
-import { frustrationsSchema } from './extract-frustrations';
+import { frustrationsSchema, memeTemplateSchema } from '../schemas';
 
-// Schema for meme template
-const memeTemplateSchema = z.object({
-  id: z.string(),
-  name: z.string(),
-  url: z.string(),
-  width: z.number(),
-  height: z.number(),
-  box_count: z.number()
-});
-
-// Type for frustrations data
-type FrustrationsData = z.infer<typeof frustrationsSchema>;
-
-export const findBaseMemeTools = createTool({
+export const findBaseMemeStep = createStep({
   id: "find-base-meme",
   description: "Search for appropriate meme templates based on user frustrations using Imgflip's free API",
-  inputSchema: z.object({
-    frustrations: frustrationsSchema,
-    style: z.string().optional().describe("Preferred meme style")
+  inputSchema: frustrationsSchema.extend({
+    analysis: z.object({
+      message: z.string()
+    })
   }),
-  execute: async ({ context: { frustrations, style } }) => {
+  outputSchema: z.object({
+    templates: z.array(memeTemplateSchema),
+    searchCriteria: z.object({
+      categories: z.array(z.string()),
+      mood: z.string(),
+      style: z.string()
+    }),
+    totalAvailable: z.number(),
+    matchingStrategy: z.string(),
+    analysis: z.object({
+      message: z.string()
+    })
+  }),
+  execute: async ({ inputData }) => {
     try {
       console.log("🔍 Searching for the perfect meme templates...");
       
@@ -38,37 +39,30 @@ export const findBaseMemeTools = createTool({
       console.log(`📚 Found ${allMemes.length} available meme templates`);
       
       // Enhanced mapping of frustration categories to relevant meme templates
-      // Using actual popular meme names from Imgflip
       const categoryMemeMap: Record<string, string[]> = {
         meetings: [
           'Distracted Boyfriend', 'Drake Pointing', 'This Is Fine', 'Boardroom Meeting Suggestion',
-          'Expanding Brain', 'Two Buttons', 'Change My Mind', 'Disaster Girl',
-          'Hide the Pain Harold', 'Awkward Moment Sealion'
+          'Expanding Brain', 'Two Buttons', 'Change My Mind', 'Disaster Girl'
         ],
         processes: [
           'This Is Fine', 'Expanding Brain', 'Drake Pointing', 'Change My Mind',
-          'Disaster Girl', 'Confused Screaming', 'Picard Facepalm', 'Y U No',
-          'First World Problems', 'Frustrated Boromir'
+          'Disaster Girl', 'Confused Screaming', 'Picard Facepalm', 'Y U No'
         ],
         technology: [
           'This Is Fine', 'Drake Pointing', 'Expanding Brain', 'Confused Screaming',
-          'Y U No', 'Picard Facepalm', 'Disaster Girl', 'Programmer Humor',
-          'Success Kid', 'Frustrated Boromir'
+          'Y U No', 'Picard Facepalm', 'Disaster Girl', 'Success Kid'
         ],
         communication: [
           'Distracted Boyfriend', 'Drake Pointing', 'Two Buttons', 'Change My Mind',
-          'Confused Screaming', 'Picard Facepalm', 'Y U No', 'Awkward Moment Sealion',
-          'Hide the Pain Harold', 'First World Problems'
+          'Confused Screaming', 'Picard Facepalm', 'Y U No', 'Hide the Pain Harold'
         ],
         management: [
           'Boardroom Meeting Suggestion', 'Drake Pointing', 'Distracted Boyfriend', 'This Is Fine',
-          'Change My Mind', 'Picard Facepalm', 'Disaster Girl', 'Y U No',
-          'Hide the Pain Harold', 'Frustrated Boromir'
+          'Change My Mind', 'Picard Facepalm', 'Disaster Girl', 'Y U No'
         ],
         workload: [
           'This Is Fine', 'Drake Pointing', 'Expanding Brain', 'Disaster Girl',
-          'Confused Screaming', 'Y U No', 'Picard Facepalm', 'First World Problems',
-          'Hide the Pain Harold', 'Frustrated Boromir'
+          'Confused Screaming', 'Y U No', 'Picard Facepalm', 'Hide the Pain Harold'
         ],
         other: [
           'Drake Pointing', 'This Is Fine', 'Distracted Boyfriend', 'Change My Mind',
@@ -78,12 +72,12 @@ export const findBaseMemeTools = createTool({
       
       // Get relevant meme names based on frustration categories
       const relevantMemeNames = new Set<string>();
-      frustrations.frustrations.forEach((frustration) => {
+      inputData.frustrations.forEach((frustration) => {
         const categoryMemes = categoryMemeMap[frustration.category] || categoryMemeMap.other;
         categoryMemes.forEach(name => relevantMemeNames.add(name.toLowerCase()));
       });
       
-      console.log(`🎯 Targeting memes for categories: ${frustrations.frustrations.map(f => f.category).join(', ')}`);
+      console.log(`🎯 Targeting memes for categories: ${inputData.frustrations.map(f => f.category).join(', ')}`);
       
       // Enhanced filtering with fuzzy matching
       let filteredMemes = allMemes.filter((meme: any) => {
@@ -97,7 +91,7 @@ export const findBaseMemeTools = createTool({
         }
         
         // Keyword matching from frustrations
-        const allKeywords = frustrations.frustrations.flatMap(f => f.keywords).map(k => k.toLowerCase());
+        const allKeywords = inputData.frustrations.flatMap(f => f.keywords).map(k => k.toLowerCase());
         return allKeywords.some(keyword => memeName.includes(keyword));
       });
       
@@ -107,15 +101,15 @@ export const findBaseMemeTools = createTool({
       if (filteredMemes.length === 0) {
         console.log("🔄 No direct matches found, searching by mood...");
         const moodMemeMap: Record<string, string[]> = {
-          frustrated: ['This Is Fine', 'Y U No', 'Frustrated Boromir', 'Picard Facepalm'],
-          annoyed: ['Drake Pointing', 'Y U No', 'Confused Screaming', 'Hide the Pain Harold'],
-          overwhelmed: ['This Is Fine', 'Disaster Girl', 'Confused Screaming', 'First World Problems'],
-          tired: ['Hide the Pain Harold', 'This Is Fine', 'Picard Facepalm', 'First World Problems'],
-          angry: ['Y U No', 'Frustrated Boromir', 'Disaster Girl', 'Confused Screaming'],
-          sarcastic: ['Drake Pointing', 'Distracted Boyfriend', 'Change My Mind', 'Success Kid']
+          frustrated: ['This Is Fine', 'Y U No', 'Picard Facepalm'],
+          annoyed: ['Drake Pointing', 'Y U No', 'Confused Screaming'],
+          overwhelmed: ['This Is Fine', 'Disaster Girl', 'Confused Screaming'],
+          tired: ['Hide the Pain Harold', 'This Is Fine', 'Picard Facepalm'],
+          angry: ['Y U No', 'Disaster Girl', 'Confused Screaming'],
+          sarcastic: ['Drake Pointing', 'Distracted Boyfriend', 'Change My Mind']
         };
         
-        const moodMemes = moodMemeMap[frustrations.overallMood] || moodMemeMap.frustrated;
+        const moodMemes = moodMemeMap[inputData.overallMood] || moodMemeMap.frustrated;
         filteredMemes = allMemes.filter((meme: any) => 
           moodMemes.some(moodMeme => 
             meme.name.toLowerCase().includes(moodMeme.toLowerCase())
@@ -143,12 +137,12 @@ export const findBaseMemeTools = createTool({
       
       console.log(`✅ Selected ${selectedMemes.length} meme templates`);
       
-      const searchResults = {
+      return {
         templates: selectedMemes,
         searchCriteria: {
-          categories: frustrations.frustrations.map((f) => f.category),
-          mood: frustrations.overallMood,
-          style: style || frustrations.suggestedMemeStyle
+          categories: inputData.frustrations.map((f) => f.category),
+          mood: inputData.overallMood,
+          style: inputData.suggestedMemeStyle
         },
         totalAvailable: allMemes.length,
         matchingStrategy,
@@ -156,18 +150,9 @@ export const findBaseMemeTools = createTool({
           message: `Found ${selectedMemes.length} great templates! Top pick: "${selectedMemes[0]?.name}"`
         }
       };
-      
-      return {
-        success: true,
-        data: searchResults
-      };
     } catch (error) {
       console.error('Error finding base memes:', error);
-      return {
-        success: false,
-        error: 'Failed to search for meme templates',
-        data: null
-      };
+      throw new Error('Failed to search for meme templates');
     }
   }
 });
