+++ src/mastra/workflows/steps/index.ts
@@ -1,0 +1,6 @@
+// Export all steps from a central location
+export { extractFrustrationsStep } from './extract-frustrations';
+export { findBaseMemeStep } from './find-base-meme';
+export { generateCaptionsStep } from './generate-captions';
+export { generateMemeStep } from './generate-meme';
+export { publishMemeStep } from './publish-meme';
